Do not attempt to switch a ruby version
Installing a new ruby version before test caused the issue below:

    Ignoring gem-wrappers-1.2.7 because its extensions are not built.  Try: gem pristine gem-wrappers --version 1.2.7
    Error loading RubyGems plugin "/opt/circleci/.rvm/gems/ruby-2.3.1/gems/gem-wrappers-1.2.7/lib/rubygems_plugin.rb": cannot load such file -- gem-wrappers (LoadError)

The root cause was unknown, but seems to be related to the "rvm install ..." command that runs in the dependencies phase. There is no reason why we shouldn't use the default ruby, so let's not switch the version and use the default one.