Fixed a bug where user can't upload photos from JSON page editor
The version of blueimp-file-upload has been upgraded unexpectedly and it seems that the version we are currently using (9.14.0) is not compatible with our js setup. This commit downgrades the library to the latest version that is compatible.

The change that broke our feature is this: https://github.com/blueimp/jQuery-File-Upload/commit/c79a44d56867771f0d420cbe4d10bc4abb08e045, which supposedly addd support for Node/CommonJS. However, it goes into the Node/CommonJS logic when we need to load it to the browser. We were not sure what the proper fix would be for this, so we decided to downgrade for the time being.

We will submit a pull request to the upstream repository to get feedback.

Signed-off-by: Yuki Nishijima <yuki@artsymail.com>