Add stub data and skeleton components
These are the results of "mobbing" with @anandaroop and @oxaudo.
We have some new fake data and beautiful orange and pink divs. One
thing Roop and I discussed was how to pass messages from the sidebar
to the content pane. In the past I've used React Router for this.
Roop mentioned Waypoints might possibly be relevant also.