Revamp Intro / StyledIntro
This sets up a pattern where we have a base component (Intro) which
produces markup that's vanilla, structural, semantic.

And we create a styled component (StyledIntro) from that, for use
in a particular context (our actual App), that applies styles
to the vanilla markup.

This should make it easier to keep the base components re-usable
and enforces a clearer separation between structure and appearance.