Fix scripts not exiting properly in CI; refactor (#2638)
* Fix incorrect shebangs in shell scripts

* Use consistent `/bin/bash` in header for running scripts

While using `/usr/bin/env bash` might be more portable, I would prefer
not take security risks. If we run scripts as executables, not found
/bin/bash will break things early and we can fix.

* Use consistent shell set builtin in scripts

It seems like the -e and -x are portable in other shells. Let's move it
out of the the header so it will still be set when the scripts are run
by a shell command.

* Always run shell scripts as executables

* Fix syntax error in deploy script