Start app in dev mode by default unless NODE_ENV is production
Previously we started the app by default in production mode, and only
started it in dev mode when $NODE_ENV is set to "development". However,
this requires us to grep NODE_ENV from .env and export it in development
before being able to perform the check.

However, in production, lacking the .env file accidentally turns the
export statement into just `export` and that means listing all the
exported variables. This not only pollutes the logs but also has
security concerns.

This commit changes the default to be development and only run the app
in production mode when $NODE_ENV is set to "production". This
simplifies the check because $NODE_ENV will be set properly in production.