Remove unused MarketInsights components for Artwork page
This commit removes a reference to a Market Insights component in
reaction directly inserted via Force.

This component does not seem to be in-use currently. Instead, it looks
like we're inserting a newer Market Insights component which is part of
the new Responsive Artwork Page in artsy/reaction.

This also unblocks removal of the older component in artsy/reaction.