Consider redirect_uri when redirecting after authentication.
TODO:

-[ ] Add test coverage to the `handleSubmit` function for this change.

Problem:

We want to be able to log into Gravity when we log into force. However,
sometimes we login to force via AJAX requests. When we do this, we need
the client to be responsible for redirecting, but the client needs to
know which URL to redirect to.

Solution:

* Patch `handleSubmit` such that it first checks if the response (from
artsy-passport on the server) has a key `redirect_uri`. If it does,
redirect to it.
  + this change is related to https://github.com/artsy/artsy-passport/pull/133
    which exposes a request URI.

Jira: https://artsyproduct.atlassian.net/browse/AUCT-281