Display the minimumNextBid in dollars if there isn’t a max bid.
* s/next/nimimumNexBidDollars for clarity

Problem:

As result of the changes in this PR, if the `bid` query param isn't set
to an integer, then the text in the bid button on mobile would display
as empty.

Solution:

Display the value of the minimum next bid in dollars as a fallback for
when there isn't a displayable value for the `bid` query parameter.

I attempted to start extracting these basic calculations and data
lookups into the routes so that they could be tested, but ran into a bit
of a rabbit hole, so backed out that for now.