add vat number display to old version of partner page
More context: this is a followup from Force #4322, which added VAT display to the newer version of the partner page, partner2. At the time, I believe we thought we didn't need to also implement it on the old version of the page (partner). As we were testing with a newly-created partner on staging, we discovered that at least that partner still did use partner instead of partner2. I would like to follow up by deprecating partner, but for now, putting up this patch.