Block a production deploy when there's a unresolved deploy block.
Problem

Right now, we can and [do][0] create deploy block records for Force,
but currently there's no systematic enforcement of preventing a
unintended deploy when there's an open unresolved deploy block.

Solution

Block a deploy using the Horizon orb. This is similar to [setup in
artsy/currents][1] that has been running production for a while.

[0]:https://releases.artsy.net/admin/deploy_blocks?utf8=%E2%9C%93&q%5Bproject_id_eq%5D=11&commit=Filter&order=id_desc
[1]:https://github.com/artsy/currents/pull/257/files