Expose stitch context to user settings template client-side
Stitch is available during server-side renders but when switching
between settings tabs in a browser which triggers a client-side render
of the same template, any stitch components fail to render due to the
missing stitch context.

I took a look at the client-side setup for the payments tab which also
uses a stitched component from artsy/reaction and found what was
missing. Pass in stitch explicitly to the template during the
client-side template render.

h/t to @eessex for catching this!