Move CSRF_TOKEN to Cookie to Allow for Page Caching
Problem

Storing the CSRF_TOKEN in the page causes issues when doing full html
caching because many users will see the same token and that token will
only be valid for a single users session.

Solution

Move the delivery of the CSRF_TOKEN from sharify to the cookie. This
allows the page to be cached while using dynamic tokens.