Fix infinite scrolling on mobile partner profile artworks tab
The issue was we tried to append fetched artworks to non-existing
element when scrolling. This quickly fixes it by appending to individual
column.

Note that it naively appends grouped artworks to column with the same
index, so it's not guaranteed the height of the columns would be
balanced. We can follow up separately if necessary.