Use Express route matching in userRequiredMiddleware
The current implementation only uses the top level path for checking if
a route should require a logged-in user or not. This makes it difficult
to allow different requirements for nested routes. For example, we might
require a logged-in user for /user/conversations but might not require
that for /user/login.

This commit switches to use regex for matching routes, and it's using
`path-to-regexp`, the same package Express.js uses for matching routes.
Therefore, we can use the same regex syntax.`