Do not animate when scrolling to bottom on initial render
Smooth scrolling transition works well when adding a message. On initial
render, we want to start from bottom as early as possible, so let's not
animate.

A better/CSS solution to start from bottom would be using
`flex-direction: column-reverse`. However, Firefox currently has a bug
that doesn't support overflow when using `*-reverse`, so there would be
no scrollbar. We currently have 3%-4% users on Firefox so let's stick
with the JavaScript approach for now.

https://caniuse.com/#search=flex-direction
https://bugzil.la/1042151