Description
Throwing an exception is a really expensive operation as it must walk
the callstack to generate a stack trace. This should be done with care
and only when nescesary.

This minor tweak increases performance slightly by only throwing when
nescesary and updating the route link matching to handle nulls instead
of relying on throwing an exception.