Tree Shaking
Description

Due to a combination of current javascript bundling restrictions, many of
which will be remedied as es modules and webpack evolve. The
`@artsy/palette` library is being fully include on every page load. This
change updates the imports so that only the required files are imported.
Thus decreasing the bundle sizes and slightly increasing page load
times.