Fix Flaky Tests
The fair_info test was causing the majority of the mocha test failures
in CI, this appears to have been caused by a deferred that was not being
waited upon.

During testing I found a second test that was flaky in rare cases where
it ran slow enough to cause a the unix timestamp to become off by a few
ms.

This change fixes these two flaky tests and should help alleviate the
retries that have become nescesary on CI.