fix(FX-2324): Move scrollY to previous position after content change
This doesn't seem to affect all browsers but, on some versions of
Chrome, the position is lost after a visitor clicks "Load more" and
content is placed above the current scroll position.

To avoid the loss of position when new content loads in, snap back to
the previous position if we're below where we were due to content
appearing above the current scroll position.