fix(fair organizer): sort fairs by -start_at to determine upcoming show
The default fair sort in Gravity is -created_at. When looking for an upcoming
fair for an organizer, we currently fetch the first fair and check if the start
time is in the future. If a past fair is created after a future fair, we might
be checking the wrong entity in the system.

This updates the GraphQL query to sort fair results by `-start_at` to mimic the
behavior in the previous implementation [here][commit].

[commit]: https://github.com/artsy/force/blob/4de69dfc0d2d26cbe43b7e5ffefc0a88e44333ac/src/desktop/apps/fair_organizer/routes.coffee#L36-L41