feature: Dockerfile Tuning
Tuns the way that caches are used in the Dockerfile.

The first change separates the dev and prod caches this fixes a rare
issues caused when the cache is empty and both instances of yarn attempt
to write the same files at the same time. This will increase total disk
space in CI by about 2 GB.

The second change limits a single writer to the cypress directory at a
time.