Provide button to copy recovery codes when enabling 2FA
Once a user enables 2FA, a modal is displayed listing the 2FA recovery
codes that can be used in the lack of the selected 2FA method.

We want to enforce the users to copy these codes, so we're adding a
button at this point that copies the list of codes to clipboard.