Hide ShippingInfo for inquireable artworks
This is a quick fix just because MOOAEA project was adding the shipping
info to artwork page in artworks that didn't contain this information
until now. This is leading to unexpected message to be displayed (like
`Shipping null`). This was happening because this message is displayed
to offerable artworks.

Changes here just aim to prevent this from being displayed in a scenario
that was affected by MOOAEA project.

The actual fix might consist on understanding why the message in being
returned like this from metaphysics. For details see this slack thread:
https://artsy.slack.com/archives/C9YNS4X32/p1644914386896509