feat: update NFTs nav link to /nft
This commit updates the NFTs nav link to a relative `/nft` href. This will
match a currently registered vanity shortcut to `/gene/nft`.

The gene is currently unpublished so the page a user is redirected to throws a
Forbidden 403 error.

We can either:
- Publish the gene, which will make the gene page visible
- Update the vanity shortcut to another visible page. The collect page filtered
  to artworks of the NFT medium type, for example.

Either way, changing the nav link to not point directly to a specific subdomain
gives us more flexibility and a UI to change where this link redirects a user
to.