fix: increase order app bottom margin
This commit increases the margin on the bottom of the container for the "interactive" (left-hand side) of the order app across all pages in the flow. This makes sure that a user can always scroll the "continue" button into view, even if they have not accepted cookies and the cookie consent banner remains stuck to the bottom of the screen.