fix(NX-3190): fix bug on CBN flow for edition sets
[NX-3190]

This fix is needed because otherwise Conversational Buy Now Flow for
artworks containing more than 1 edition set would lead to "Make Offer"
flow instead of "Buy Now".

The solution here consists of passing the param between different
components so we can control if an order is an offer or purchase. We
understand this is not the ideal solution, this is a quick fix so we can
release the feature tomorrow. We'll work separately on a proper refactor
to avoid the prop drilling.