fix: avoid triggering address verification modal again after choosing an address
The address verification modal gets triggered again after user choosing
an address from it. Setting states is async. Checking if the
verification needs to run right after the async set states will get the
old value of the states, which triggers the address verification again.

Since on choosing an address we are ready to proceed to the next step,
this directly move on to finalize fulfillment, instead of checking if
address verification needs to run.