fix: display set payment by setup intent error correctly
Checking `!isPaymentSetupSuccessful` is not sufficient to display the
error because `isPaymentSetupSuccessful` is default false when setting
payment is not yet performed. This causes the error modal to be display
incorrectly when the page first loads.

This replaces `paymentSetupErrorCode` with `paymentSetupError` and use
it to determine showing the error modal or not.