build: relax rule about `describe` names
We've had a long-standing (but not universal) pattern of using
actual function references as the first argument to `describe` blocks,
as opposed to strings that correspond to the function name.

This is handy because TS prevents these from getting out of sync, in the
event of refactors or renames. I think we should consider this
a feature and not a bug.

Unfortunately this rule currently nudges us to undo this pattern, even
though there has been an escape hatch from this for quite a while
https://github.com/jest-community/eslint-plugin-jest/pull/443/commits/838f0be