feat(artist series filter): add to artist page's artwork grid (#13041)
* build: relax rule about `describe` names

We've had a long-standing (but not universal) pattern of using
actual function references as the first argument to `describe` blocks,
as opposed to strings that correspond to the function name.

This is handy because TS prevents these from getting out of sync, in the
event of refactors or renames. I think we should consider this
a feature and not a bug.

Unfortunately this rule currently nudges us to undo this pattern, even
though there has been an escape hatch from this for quite a while
https://github.com/jest-community/eslint-plugin-jest/pull/443/commits/838f0be

* feat: support filtering by multiple OR'd artist series

* feat: support aggregation on artist series

* feat: create Artist Series facet filter

* feat: wire up new filter to artist artwork grid

* feat: enable display of selected artist series as pills

* feat: hide the new filter behind a feature flag

* perf: only request artist series agg if feature is enabled

…and a small ocd naming consistency fix