chore: cleanup unused default alert placeholder value (#13073)
Follow-up to [#13059][].

As of [#12751][], we no longer rely on an explicitly passed value for an
alert's default name, also rendered as a placeholder on the text input. Instead, we query Metaphysics for a default/generated name based on the working alert criteria, which we also show in the input placeholder but no longer persist with the alert. This default/generated name is also used after the alert has been created and stays in sync with any changes to the criteria.

The previous cleanup PR ([#13059][]) removed the feature flag branching logic
and this PR follow-ups and removes the `placeholder` attribute from the
`SavedSearchEntity` type and any instances of a placeholder passed in
elsewhere.

[#13059]: https://github.com/artsy/force/pull/13059
[#12751]: https://github.com/artsy/force/pull/12751