feat: allow truncation in alert options UI
Artist Series list may end up very long, so we want to ensure that
we can display a smaller subset for starters.

This Show More behavior was previously abandoned (see #13106 and #13156)
but is now necessary again. So I have taken another swing at it,
this time trying to address some of the layout issues that prevented
the use of the Palette ShowMore component.