Fix Salesforce Chat Bubble for auctions (#13570)
The current implementation is not working on the auction page because it
needs different handling for the price thresold.

The chat bubble for artworks is based on the listed price. This works
for BNMO artworks, but it doesn't work for Auction artworks. For Auction
artworks we want to consider the starting bid value instead of listed
price.