feat: defer forced login / sign up to end of SWA submission flow
Instead of forcing a user to login or sign up directly on the /sell landing
page, we'd like to allow the user to start and progress through the submission
and only force an authenticated state at the end, right before the submission
graduates from draft > submitted. At this point, we want to make sure we have
an Artsy user account to link the submission but we're okay with letting vistor
/ unauthenticated users start and experience the submission flow before that
final step.