Update the class name on css to match the component (#15016)
We must not display the SF chat bubble on the mobile version. We handled
this during the migration from Zendesk to SF, but our css rule specifies
a class that's defined on SF stylesheet, which became outdated after the
recent upgrade.

Changes in this PR point update the class name to meet the current
definition.